grsec: Apply grsecurity-2.1.14-2.6.32.8-201002101507.patch
Signature-tree: b9d728edaa4ff1e6c309729a5003d5992cc1b1c1

+++ arch/x86/mm/fault.c
@@ -1508,7 +1508,7 @@ void pax_report_insns(void *pc, void *sp)
 	printk("\n");
 
 	printk(KERN_ERR "PAX: bytes at SP-%lu: ", (unsigned long)sizeof(long));
-	for (i = -1; i < 80 / sizeof(long); i++) {
+	for (i = -1; i < 80 / (long)sizeof(long); i++) {
 		unsigned long c;
 		if (get_user(c, (__force unsigned long __user *)sp+i))
 #ifdef CONFIG_X86_32
